app/requirements.txt
flask
azure.identity
azure.keyvault.secrets
pyodbc